+++ src/dimeval.hpp
@@ -62,10 +62,10 @@ namespace dv {
         UnitVector operator^(const UnitVector &rhs) const noexcept;
     };
     struct EValue {
-        double value;
+        long double value;
         UnitVector unit;
-        EValue(const double val): value{val}, unit{DIMENSIONLESS_VEC} {}
-        EValue(const double val, const UnitVector _unit): value{val}, unit{_unit} {}
+        EValue(const long double val): value{val}, unit{DIMENSIONLESS_VEC} {}
+        EValue(const long double val, const UnitVector _unit): value{val}, unit{_unit} {}
         EValue operator+() const noexcept;
         EValue operator+(const EValue &rhs) const noexcept;
         EValue operator-() const noexcept;
